+++ lib/acts_as_tenant/errors.rb
@@ -9,9 +9,6 @@ module ActsAsTenant
     class NoTenantSet < ActsAsTenant::Error
     end
 
-    class ModelNotScopedByTenant < ActsAsTenant::Error
-    end
-
     class TenantIsImmutable < ActsAsTenant::Error
     end
   end