+++ lib/acts_as_tenant/sidekiq.rb
@@ -1,6 +1,14 @@
 module ActsAsTenant::Sidekiq
+  class BaseMiddleware
+    def self.sidekiq_7_and_up?
+      Gem::Version.new(Sidekiq::VERSION) >= Gem::Version.new("7")
+    end
+  end
+
   # Get the current tenant and store in the message to be sent to Sidekiq.
-  class Client
+  class Client < BaseMiddleware
+    include Sidekiq::ClientMiddleware if sidekiq_7_and_up?
+
     def call(worker_class, msg, queue, redis_pool)
       if ActsAsTenant.current_tenant.present?
         msg["acts_as_tenant"] ||=
@@ -15,7 +23,9 @@ module ActsAsTenant::Sidekiq
   end
 
   # Pull the tenant out and run the current thread with it.
-  class Server
+  class Server < BaseMiddleware
+    include Sidekiq::ServerMiddleware if sidekiq_7_and_up?
+
     def call(worker_class, msg, queue)
       if msg.has_key?("acts_as_tenant")
         account = msg["acts_as_tenant"]["class"].constantize.find msg["acts_as_tenant"]["id"]